tinyalsa: add separate call for pcm prepare
BZ: 73509

Tinyalsa combines PREPARE and START calls to the driver in the pcm_start()
function. But in case of Aware, we do not want to call START. So have a separate
call for PREPARE.

For Probe stream, the MMX_PROBE needs to be sent after alloc, hence calling the
prepare call before MMX_PROBE call.

Change-Id: Ie31bee90dbfdca5bbd5d1a9c65239e83e365970d
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>
Reviewed-on: http://android.intel.com:8080/89943
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Vinnakota, Lakshmi N <lakshmi.n.vinnakota@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>