Mixer: Check validity of input value
BZ: 129089
[CLEAN]

When the user wants to set a mixer with a value greater
than the maximum or smaller than the minimum, the mixer
is set at a ramdom value.

This patch set purpose is to return a error if the input
value is out of range.

Change-Id: Iee7fcfccfe78ce3196b77ed1938b0f22efdd6bc3
Orig-Change-Id: I3ec4d17a8f4b4baf88b256dfd61b34f23cba8b27
Category: aosp improvement
Domain: AudioComms-Tinyalsa
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Baptiste Robert <baptistex.robert@intel.com>