Tinyalsa: standardize error message for wrong input value
BZ: 138170

Currently, if a user sets a mixer with a percentage value
greater than the maximum, tinyalsa does not inform the user
about this "error" and silently replaces the user's value
by the maximum. However, if a user sets a mixer with an
integer value greater than the maximum, tinyalsa informs
the user with an error.

This patch purpose is to define the same behaviour in both
quoted cases: user is informed through an error.

Category: aosp improvement
Domain: AudioComms-Tinyalsa
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I05290f4b2a1ca53d37ca9a73fe75c2da18c0231d
Signed-off-by: Baptiste Robert <baptistex.robert@intel.com>
Reviewed-on: http://android.intel.com:8080/131703
Reviewed-by: Wagner, David <david.wagner@intel.com>
Reviewed-by: Fomitchev, AndreiX <andreix.fomitchev@intel.com>
Reviewed-by: Pichon, SylvainX <sylvainx.pichon@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>